+++ webdna.sh
@@ -11,16 +11,20 @@ APACHE_CONFIG="${APACHE_CONFIG:/etc/httpd/conf/httpd.conf}"
  # Pull in the WebDNA folder and fcgi
 wget http://webdna.us/download/${DNA_FOLDER}.zip
 unzip -d . "${DNA_FOLDER}.zip"
-rm "${DNA_FOLDER}.zip"
+rm -f "${DNA_FOLDER}.zip"
 wget http://www.webdna.us/download/${DNA_EXECUTABLE}.zip
 unzip -d . "${DNA_EXECUTABLE}.zip"
-rm "${DNA_EXECUTABLE}.zip"
+rm -f "${DNA_EXECUTABLE}.zip"
 
 # 'Install' WebDNA
 mv "${DNA_EXECUTABLE}/WebDNA.fcgi" "WebDNA/WebDNA.fcgi"
 chmod 0755 "WebDNA/WebDNA.fcgi"
 cp -R "WebDNA" ${WEBDNA_LOC}
 chown -R $APACHE_USER:$APACHE_USER ${WEBDNA_LOC}
+
+# Remove WebDNA Artifacts
+#rm -rf "${DNA_EXECUTABLE}"
+#rm -rf "${DNA_FOLDER}"
 
 # Quick Config of Apache
 mkdir -p "$WEBROOT/public"
@@ -34,6 +38,8 @@ AddType text/html .tpl
 <Location ~  \"/.*/ErrorLog($|\.txt)\">
     deny from all
 </Location>
+<Directory "$WEBROOT">
+Options -Indexes +FollowSymLinks +ExecCGI
 FcgidWrapper ${WEBDNA_LOC}/WebDNA.fcgi .dna
 FcgidWrapper ${WEBDNA_LOC}/WebDNA.fcgi .html
 FcgidWrapper ${WEBDNA_LOC}/WebDNA.fcgi .tpl
@@ -43,6 +49,7 @@ AddHandler fcgid-script .dna .html .tpl .htm
     RewriteEngine on
     RewriteRule .* - [E=HTTP_AUTHORIZATION:%{HTTP:Authorization},L]
 </IfModule>
+</Directory>
 DefaultMaxClassProcessCount 1
 " >> "${APACHE_CONFIG}"
 # Quick Configure WebDNA