+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Docker WebDNA Changelog
 
+## (04/04/2022) v-0.2.0:
+- Added support for Ubuntu 16.04 FCGI 8.6.5
+- Optimized Webdna script
+- Optimized CentOS 7 Dockerfile
+
 ## (03/31/2022) v-0.1.0:
 - Initial Build - Added support for CentOS 7 FCGI 8.6.5